anisotropic distance map implementations

+++ src/shaders/iso_viewer/chunks/uniforms/uniforms_intensity_map.glsl
@@ -3,7 +3,7 @@
 
 struct IntensityMap 
 {
-    vec3  dimensions;    
+    ivec3  dimensions;    
     vec3  spacing;           
     vec3  size;         
     float spacing_length;
